+++ CotEditor/Sources/NSAppearance.swift
@@ -8,7 +8,7 @@
 //
 //  ---------------------------------------------------------------------------
 //
-//  © 2018-2023 1024jp
+//  © 2018-2024 1024jp
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
 //  you may not use this file except in compliance with the License.
@@ -43,6 +43,75 @@ extension NSAppearance {
                 true
             default:
                 false
+        }
+    }
+    
+    
+    /// THe corresponding appearance for the light/dark mode.
+    ///
+    /// - Parameter isDark: Whether the return value to be for dark mode or not.
+    /// - Returns: A new appearance, or `self` if no change required.
+    final func appearance(for isDark: Bool) -> Self {
+        
+        if isDark == self.isDark {
+            self
+        } else {
+            Self(named: self.name.name(for: isDark)) ?? self
+        }
+    }
+}
+
+
+extension NSColor {
+    
+    /// Forcibly creates the appropriate color for the given appearance.
+    ///
+    /// - Parameter appearance: The appearance to match.
+    /// - Returns: A new color.
+    func solve(for appearance: NSAppearance?) -> NSColor {
+        
+        guard self.type != .componentBased else { return self }
+        
+        var color = self
+        appearance?.performAsCurrentDrawingAppearance {
+            color = NSColor(cgColor: color.cgColor) ?? color
+        }
+        return color
+    }
+    
+    
+    /// Forcibly creates the appropriate color for the light/dark mode.
+    ///
+    /// - Parameter isDark: Whether the return value to be for dark mode or not.
+    /// - Returns: A new color.
+    func forDarkMode(_ isDark: Bool) -> NSColor {
+        
+        guard self.type != .componentBased else { return self }
+        
+        return self.solve(for: NSAppearance(named: isDark ? .darkAqua : .aqua)!)
+    }
+}
+
+
+private extension NSAppearance.Name {
+    
+    /// The corresponding appearance name for the light/dark mode.
+    ///
+    /// - Parameter isDark: Whether the return value to be for dark mode or not.
+    /// - Returns: An appearance name.
+    func name(for isDark: Bool) -> Self {
+        
+        switch self {
+            case .aqua, .darkAqua:
+                isDark ? .darkAqua : .aqua
+            case .vibrantLight, .vibrantDark:
+                isDark ? .vibrantDark : .vibrantLight
+            case .accessibilityHighContrastAqua, .accessibilityHighContrastDarkAqua:
+                isDark ? .accessibilityHighContrastDarkAqua : .accessibilityHighContrastAqua
+            case .accessibilityHighContrastVibrantLight, .accessibilityHighContrastVibrantDark:
+                isDark ? .accessibilityHighContrastVibrantDark : .accessibilityHighContrastVibrantLight
+            default:
+                self
         }
     }
 }