+++ Tests/ThemeTests.swift
@@ -46,7 +46,7 @@ final class ThemeTests: XCTestCase {
         XCTAssertEqual(theme.invisibles.color.brightnessComponent, 0.72, accuracy: 0.01)
         XCTAssertEqual(theme.background.color, NSColor.white.usingColorSpace(.genericRGB))
         XCTAssertEqual(theme.lineHighlight.color.brightnessComponent, 0.93, accuracy: 0.01)
-        XCTAssertNil(theme.secondarySelectionColor)
+        XCTAssertEqual(theme.effectiveSecondarySelectionColor(for: NSAppearance(named: .aqua)!), .unemphasizedSelectedContentBackgroundColor)
         XCTAssertFalse(theme.isDarkTheme)
         
         for type in SyntaxType.allCases {